+++ Plutarch/Internal/TryFrom.hs
@@ -81,7 +81,7 @@ class PSubtype a b => PTryFrom (a :: S -> Type) (b :: S -> Type) where
 ptryFrom :: forall b a s r. PTryFrom a b => Term s a -> ((Term s b, Reduce (PTryFromExcess a b s)) -> Term s r) -> Term s r
 ptryFrom = ptryFrom'
 
-newtype Flip f a b = Flip (f b a) deriving stock (Generic)
+newtype Flip (f :: k1 -> k2 -> Type) (a :: k2) (b :: k1) = Flip (f b a) deriving stock (Generic)
 
 instance PTryFrom PData (PAsData PInteger) where
   type PTryFromExcess PData (PAsData PInteger) = Flip Term PInteger