+++ plutarch-ledger-api/src/Plutarch/LedgerApi/V1/Credential.hs
@@ -8,6 +8,7 @@ module Plutarch.LedgerApi.V1.Credential (
 
 import GHC.Generics (Generic)
 import Generics.SOP qualified as SOP
+import Plutarch.Internal.Parse (PValidateData)
 import Plutarch.LedgerApi.V1.Crypto (PPubKeyHash)
 import Plutarch.LedgerApi.V1.Scripts (PScriptHash)
 import Plutarch.Prelude
@@ -38,6 +39,8 @@ data PCredential (s :: S)
   deriving
     ( -- | @since 3.3.0
       PlutusType
+    , -- | @since 3.5.0
+      PValidateData
     )
     via (DeriveAsDataStruct PCredential)
 
@@ -74,6 +77,8 @@ data PStakingCredential (s :: S)
   deriving
     ( -- | @since 3.3.0
       PlutusType
+    , -- | @since 3.5.0
+      PValidateData
     )
     via (DeriveAsDataStruct PStakingCredential)